[Edl98PLU] Fix comments
